[INFO] - Recover anime info from file name
[-] Remove yesterday code
[+] Adding  a new open source librairy which extract anime info from filename
    ---[+] Adding librairy to requirements.txt : anitopy>=2.1.0
    ---[*] Replace INT to FLOAT for episode collumns in episode_info db
    ---[*] Update README.md file

+++ src/database/episode_info.py
@@ -25,7 +25,7 @@ episode_psql_creation = '''CREATE TABLE episode_info (
                                 id SERIAL NOT NULL,
                                 anime_name VARCHAR(100) NOT NULL,
                                 saison INT NOT NULL ,
-                                episode INT NOT NULL ,
+                                episode FLOAT NOT NULL ,
                                 clip_duration INT NOT NULL ,
                                 hash VARCHAR(42) NOT NULL,
                                 done_at INT NULL,
